Support for Clang compiler under MacOSX
Boost compiled using:
./b2 link=static --without-python runtime-link=static toolset=clang
cxxflags="-stdlib=libc++" linkflags="-stdlib=libc++"

ShadingZenCpp compiled using:
./dependencies/boost/b2 link=static

+++ tests/Math/matrix4f_tests.cpp
@@ -1,7 +1,7 @@
 #include <boost/test/unit_test.hpp>
 #include <shz/math/matrix.hpp>
  
-extern bool is_aligned(void *p, int N);
+extern bool is_aligned(void *p, intptr_t N);
 
 BOOST_AUTO_TEST_CASE(matrix4fConstructors)
 {
